requirements.txt
Django
pytz
requests
djangorestframework
django-filter
psycopg2-binary
python-dotenv
drf-extra-fields
django-rest-passwordreset
PyYAML
ujson
gunicorn
redis
celery
drf-spectacular
